Updated Youtube playback cases and few modifications in other test cases
