+++ src/tests/youtube/Cobalt_Visibility_001.test.js
@@ -6,33 +6,30 @@ import {
   suspendOrResumeCobaltPlugin,
 } from '../../commonMethods/cobalt'
 
+let listener
 export default {
   title: 'Cobalt Visibility - 001',
   description: 'Set Cobalt Visibility to Hidden and check the visibility state',
   context: {
     visibilityState: 'hidden',
   },
+  setup() {
+    return this.$sequence([
+      () => pluginDeactivate.call(this, constants.youTubePlugin),
+      () => pluginActivate.call(this, constants.youTubePlugin),
+      () => suspendOrResumeCobaltPlugin.call(this, constants.resume),
+      () =>
+        (listener = this.$thunder.api.Cobalt.on('visibilitychange', data => {
+          this.$data.write('visibility', data.hidden)
+        })),
+    ])
+  },
+  teardown() {
+    listener.dispose()
+  },
   steps: [
     {
-      description: 'Deactivate Cobalt Plugin',
-      test: pluginDeactivate,
-      params: constants.youTubePlugin,
-      assert: 'deactivated',
-    },
-    {
-      description: 'Activate Youtube Plugin and check suspended or not',
-      test: pluginActivate,
-      params: constants.youTubePlugin,
-      assert: 'suspended',
-    },
-    {
-      description: 'Resume Cobalt Plugin and check resumed or not',
-      test() {
-        suspendOrResumeCobaltPlugin.call(this, constants.resume)
-      },
-    },
-    {
-      description: 'Set Cobalt Plugin visibility',
+      description: 'Set Cobalt Browser visibility',
       test() {
         return setCobaltVisibility.call(this, this.$context.read('visibilityState'))
       },
@@ -40,8 +37,27 @@ export default {
         if (res === null) {
           return true
         } else {
-          throw new Error('Result is not as expected')
+          throw new Error('Proper error message is not shown')
         }
+      },
+    },
+    {
+      description: 'Wait until visiblity change event is changed',
+      sleep() {
+        // Purpose of this sleep is to wait until current step gets 'visibility change' response from the listener
+        return new Promise((resolve, reject) => {
+          let attempts = 0
+          const interval = setInterval(() => {
+            attempts++
+            if (this.$data.read('visibility') === true) {
+              clearInterval(interval)
+              resolve()
+            } else if (attempts > 10) {
+              clearInterval(interval)
+              reject('Visibility not changed')
+            }
+          }, 1000)
+        })
       },
     },
     {
@@ -53,7 +69,7 @@ export default {
         if (res === this.$context.read('visibilityState')) {
           return true
         } else {
-          throw new Error('Visibility is not as expected and is ', res)
+          throw new Error(`Visibility is not as expected and is ${res}`)
         }
       },
     },
